refactor: make ocean-heat-content simpler
