+++ examples/ocean-heat-content/ocean-heat-content-1972-03-2000m/ocean-heat-content-1972-03-2000m.json
@@ -5,7 +5,7 @@
   "properties": {
     "start_datetime": "1972-03-01T00:00:00Z",
     "end_datetime": "1972-03-31T23:59:59Z",
-    "created": "2022-09-01T10:01:30.081329Z",
+    "created": "2022-09-28T11:49:03.321445Z",
     "proj:epsg": 4326,
     "proj:shape": [
       180,
