+++ examples/ocean-heat-content/ocean-heat-content-2017-2021-100m/ocean-heat-content-2017-2021-100m.json
@@ -5,7 +5,7 @@
   "properties": {
     "start_datetime": "2017-01-01T00:00:00Z",
     "end_datetime": "2021-12-31T23:59:59Z",
-    "created": "2022-09-01T10:01:31.193434Z",
+    "created": "2022-09-28T11:49:04.529476Z",
     "proj:epsg": 4326,
     "proj:shape": [
       180,
