+++ examples/ocean-heat-content/ocean-heat-content-2017-2021-700m/ocean-heat-content-2017-2021-700m.json
@@ -5,7 +5,7 @@
   "properties": {
     "start_datetime": "2017-01-01T00:00:00Z",
     "end_datetime": "2021-12-31T23:59:59Z",
-    "created": "2022-09-01T10:01:29.705862Z",
+    "created": "2022-09-28T11:49:02.909537Z",
     "proj:epsg": 4326,
     "proj:shape": [
       180,
