+++ examples/ocean-heat-content/ocean-heat-content-2021-2000m/ocean-heat-content-2021-2000m.json
@@ -5,7 +5,7 @@
   "properties": {
     "start_datetime": "2021-01-01T00:00:00Z",
     "end_datetime": "2021-12-31T23:59:59Z",
-    "created": "2022-09-01T10:01:30.420497Z",
+    "created": "2022-09-28T11:49:03.713891Z",
     "proj:epsg": 4326,
     "proj:shape": [
       180,
