+++ examples/ocean-heat-content/ocean-heat-content-2022-Q1-700m/ocean-heat-content-2022-Q1-700m.json
@@ -5,7 +5,7 @@
   "properties": {
     "start_datetime": "2022-01-01T00:00:00Z",
     "end_datetime": "2022-03-31T23:59:59Z",
-    "created": "2022-09-01T10:01:29.837148Z",
+    "created": "2022-09-28T11:49:03.072019Z",
     "proj:epsg": 4326,
     "proj:shape": [
       180,
