+++ scripts/extract_netcdf_asset_metadata.py
@@ -8,7 +8,7 @@ it on the fly.
 Progress and status are printed to stderr, so you can redirect stdout to a file.
 
 Usage:
-    scripts/extract_netcdf_asset_metadata.py > src/stactools/noaa_cdr/collection-asset-metadata.json
+    scripts/extract_netcdf_asset_metadata.py > src/stactools/noaa_cdr/ocean-heat-content.json
 """
 
 import json
@@ -20,19 +20,16 @@ import fsspec
 import xarray
 from tqdm import tqdm
 
-from stactools.noaa_cdr import Cdr
+import stactools.noaa_cdr.ocean_heat_content
 
-metadata: Dict[str, Dict[str, Dict[str, str]]] = {}
-for cdr in Cdr.cdrs():
-    print(f"Reading NetCDF files for {cdr.slug()}", file=sys.stderr)
-    metadata[cdr.slug()] = {}
-    for href in tqdm(cdr.hrefs()):
-        key = os.path.splitext(os.path.basename(href))[0]
-        with fsspec.open(href) as file:
-            with xarray.open_dataset(file, decode_times=False) as dataset:
-                metadata[cdr.slug()][key] = {
-                    "title": dataset.title,
-                    "description": dataset.summary,
-                }
+metadata: Dict[str, Dict[str, str]] = {}
+for href in tqdm(list(stactools.noaa_cdr.ocean_heat_content.noaa_hrefs())):
+    key = os.path.splitext(os.path.basename(href))[0]
+    with fsspec.open(href) as file:
+        with xarray.open_dataset(file, decode_times=False) as dataset:
+            metadata[key] = {
+                "title": dataset.title,
+                "description": dataset.summary,
+            }
 
-json.dump(metadata, sys.stdout)
+json.dump(metadata, sys.stdout, indent=4)
