--- src/stactools/noaa_cdr/cogify.py
@@ -1,123 +1,0 @@
-import copy
-import datetime
-import os.path
-from dataclasses import dataclass
-from typing import Any, Dict, Hashable, List, Optional
-
-import fsspec
-import numpy
-import rasterio
-import rasterio.shutil
-import xarray
-from pystac.extensions.raster import DataType
-from rasterio import Affine, MemoryFile
-
-from stactools.noaa_cdr import utils
-
-from .constants import EPSG, GDAL_TRANFORM, TimeResolution
-
-BASE_TIME = datetime.datetime(1955, 1, 1)
-GTIFF_PROFILE = {
-    "crs": f"epsg:{EPSG}",
-    "width": 360,
-    "height": 180,
-    "dtype": DataType.FLOAT32,
-    "nodata": numpy.nan,
-    "count": 1,
-    "transform": Affine.from_gdal(*GDAL_TRANFORM),
-    "driver": "GTiff",
-}
-COG_PROFILE = {"compress": "deflate", "blocksize": 512, "driver": "COG"}
-
-
-@dataclass(frozen=True)
-class Cog:
-    path: str
-    time_resolution: TimeResolution
-    start_datetime: datetime.datetime
-    end_datetime: datetime.datetime
-    datetime: datetime.datetime
-    epsg: int
-    shape: List[int]
-    transform: List[float]
-    nodata: float
-    data_type: DataType
-    unit: str
-    attributes: Dict[Hashable, Any]
-
-    def time_interval_as_str(self) -> str:
-        """Returns this COGs time resolution and datetime as a string.
-
-        Returns:
-            str: The time interval
-        """
-        return utils.time_interval_as_str(self.datetime, self.time_resolution)
-
-
-def cogify(
-    href: str, outdir: Optional[str] = None, latest_only: bool = False
-) -> List[Cog]:
-    """Creates a Cloud-Optimized GeoTIFF from a CDR NetCDF.
-
-    Args:
-        href (str): Input NetCDF href.
-        outdir (str, optional): Output directory for the COG. Defaults to None.
-            If None, the COG will be created alongside the input NetCDF. If href
-            is a url and outdir is not provided, an Exception is raised.
-        latest_only (bool): Only create COGs for the latest data. Useful for
-            creating a small subset of a CDRs items. Defaults to false.
-
-    Returns:
-        Cog: The created Cog.
-    """
-    if outdir is None:
-        if href.startswith("http"):
-            raise ValueError(f"Output directory is required for http hrefs: {href}")
-        outdir = os.path.dirname(href)
-    cogs = list()
-    with fsspec.open(href) as file:
-        with xarray.open_dataset(file, decode_times=False) as dataset:
-            time_resolution = TimeResolution.from_value(
-                dataset.time_coverage_resolution
-            )
-            variable = utils.data_variable_name(dataset)
-            num_records = len(dataset[variable].time)
-            for i, month_offset in enumerate(dataset[variable].time):
-                if latest_only and i < (num_records - 1):
-                    continue
-                time = utils.add_months_to_datetime(BASE_TIME, month_offset)
-                start_datetime, end_datetime = utils.datetime_bounds(
-                    time, time_resolution
-                )
-                suffix = utils.time_interval_as_str(time, time_resolution)
-                output_path = os.path.join(
-                    outdir,
-                    f"{os.path.splitext(os.path.basename(href))[0]}_{suffix}.tif",
-                )
-                values = dataset[variable].isel(time=i).values.squeeze()
-                profile = copy.deepcopy(GTIFF_PROFILE)
-                unit = dataset[variable].units.replace("_", " ")
-                profile["unit"] = unit
-                with MemoryFile() as memory_file:
-                    with memory_file.open(**GTIFF_PROFILE) as open_memory_file:
-                        open_memory_file.write(values, 1)
-                        rasterio.shutil.copy(
-                            open_memory_file, output_path, **COG_PROFILE
-                        )
-                cogs.append(
-                    Cog(
-                        output_path,
-                        time_resolution,
-                        datetime=time,
-                        start_datetime=start_datetime,
-                        end_datetime=end_datetime,
-                        epsg=EPSG,
-                        shape=[GTIFF_PROFILE["height"], GTIFF_PROFILE["width"]],
-                        transform=GDAL_TRANFORM,
-                        nodata=GTIFF_PROFILE["nodata"],
-                        data_type=GTIFF_PROFILE["dtype"],
-                        unit=unit,
-                        attributes=dataset.attrs,
-                    )
-                )
-    return cogs
