+++ tests/__init__.py
@@ -3,11 +3,11 @@ from typing import Any, Dict
 
 from stactools.testing.test_data import TestData
 
-from stactools.noaa_cdr.cdr import OceanHeatContent
+from stactools.noaa_cdr import ocean_heat_content
 
 external_data: Dict[str, Any] = dict()
 
-for href in OceanHeatContent.hrefs():
+for href in ocean_heat_content.noaa_hrefs():
     external_data[os.path.basename(href)] = {"url": href}
 
 test_data = TestData(__file__, external_data)
