--- tests/test_cogify.py
@@ -1,43 +1,0 @@
-import os.path
-from tempfile import TemporaryDirectory
-
-import pytest
-
-from stactools import noaa_cdr
-from tests import test_data
-
-COGIFY_PARAMETERS = [
-    ("heat_content_anomaly_0-700_yearly.nc", 67),
-    ("heat_content_anomaly_0-2000_monthly.nc", 207),
-    ("heat_content_anomaly_0-2000_pentad.nc", 63),
-    ("heat_content_anomaly_0-2000_seasonal.nc", 69),
-    ("heat_content_anomaly_0-2000_yearly.nc", 17),
-    ("mean_halosteric_sea_level_anomaly_0-2000_yearly.nc", 17),
-    ("mean_salinity_anomaly_0-2000_yearly.nc", 17),
-    ("mean_temperature_anomaly_0-2000_yearly.nc", 17),
-    ("mean_thermosteric_sea_level_anomaly_0-2000_yearly.nc", 17),
-    ("mean_total_steric_sea_level_anomaly_0-2000_yearly.nc", 17),
-]
-
-
-@pytest.mark.parametrize("infile,num_cogs", COGIFY_PARAMETERS)
-def test_cogify(infile: str, num_cogs: int) -> None:
-    external_data_path = test_data.get_external_data(infile)
-    with TemporaryDirectory() as temporary_directory:
-        cogs = noaa_cdr.cogify(external_data_path, temporary_directory)
-        assert len(cogs) == num_cogs
-        for cog in cogs:
-            assert os.path.exists(cog.path)
-
-
-def test_cogify_href(cogify_href: str) -> None:
-    with TemporaryDirectory() as temporary_directory:
-        cogs = noaa_cdr.cogify(cogify_href, temporary_directory)
-        assert len(cogs) == 17
-        for cog in cogs:
-            assert os.path.exists(cog.path)
-
-
-def test_cogify_href_no_output_directory(cogify_href: str) -> None:
-    with pytest.raises(Exception):
-        noaa_cdr.cogify(cogify_href)
